+++ core/src/banking_stage/transaction_scheduler/mod.rs
@@ -1,7 +1,9 @@
 mod batch_id_generator;
+pub(crate) mod greedy_scheduler;
 mod in_flight_tracker;
 pub(crate) mod prio_graph_scheduler;
 pub(crate) mod receive_and_buffer;
+pub(crate) mod scheduler;
 pub(crate) mod scheduler_controller;
 pub(crate) mod scheduler_error;
 mod scheduler_metrics;